backend/requirements.txt
feedparser==6.0.11
beautifulsoup4==4.11.1
selenium==4.18.1
flask==3.0.2
watchdog~=3.0
flask-cors==4.0.0
gunicorn==21.2.0
pymongo==4.6.2
supervisor==4.2.5
# httpx needs to be compatible with mistralai
httpx[http2]<0.26.0,>=0.25.2 
python-dotenv==1.0.1
requests==2.31.0
OpenAI==1.13.3
mistralai==0.1.3
nomic==3.0.15
